+++ models/project.go
@@ -1,7 +1,7 @@
 package models
 
 type Project struct {
-	Activitys    []Activity `gorm:"ForeignKey:ProjectID"`
+	Activitys    []Activity `json:"-" gorm:"ForeignKey:ProjectID"`
 	ID           uint
 	UserID       uint   `json:"userid"`
 	Project_name string `json:"project_name"`
