+++ models/user.go
@@ -1,10 +1,10 @@
 package models
 
 type User struct {
-	Projects  []Project  `gorm:"foreignKey:UserID"`
-	Activitys []Activity `gorm:"foreignKey:UserID"`
+	Projects  []Project  `json:"-" gorm:"foreignKey:UserID"`
+	Activitys []Activity `json:"-" gorm:"foreignKey:UserID"`
 	ID        uint
 	Username  string `json:"username"`
 	Email     string `gorm:"typevarchar(100);unique_index" json:"email"`
-	Password  string `json:"password"`
+	Password  string `json:"-"`
 }
